+++ ignite/cmd/chain.go
@@ -104,6 +104,7 @@ chain.
 		NewChainFaucet(),
 		NewChainSimulate(),
 		NewChainDebug(),
+		NewChainLint(),
 	)
 
 	return c