+++ ignite/cmd/plugin.go
@@ -18,6 +18,7 @@ import (
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
 	"github.com/ignite/cli/v29/ignite/pkg/gomodule"
 	"github.com/ignite/cli/v29/ignite/pkg/xgit"
+	"github.com/ignite/cli/v29/ignite/services/chain"
 	"github.com/ignite/cli/v29/ignite/services/plugin"
 )
 
@@ -202,15 +203,14 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			}
 		}
 
-		// Get chain when the plugin runs inside an blockchain app
-		c, err := newChainWithHomeFlags(cmd)
-		if err != nil && !errors.Is(err, gomodule.ErrGoModNotFound) {
+		api, err := newAppClientAPI(cmd)
+		if err != nil {
 			return err
 		}
 
 		ctx := cmd.Context()
 		execHook := newExecutedHook(hook, cmd, args)
-		err = p.Interface.ExecuteHookPre(ctx, execHook, plugin.NewClientAPI(plugin.WithChain(c)))
+		err = p.Interface.ExecuteHookPre(ctx, execHook, api)
 		if err != nil {
 			return errors.Errorf("app %q ExecuteHookPre() error: %w", p.Path, err)
 		}
@@ -224,15 +224,14 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			err := runCmd(cmd, args)
 			// if the command has failed the `PostRun` will not execute. here we execute the cleanup step before returnning.
 			if err != nil {
-				// Get chain when the plugin runs inside an blockchain app
-				c, err := newChainWithHomeFlags(cmd)
-				if err != nil && !errors.Is(err, gomodule.ErrGoModNotFound) {
+				api, err := newAppClientAPI(cmd)
+				if err != nil {
 					return err
 				}
 
 				ctx := cmd.Context()
 				execHook := newExecutedHook(hook, cmd, args)
-				err = p.Interface.ExecuteHookCleanUp(ctx, execHook, plugin.NewClientAPI(plugin.WithChain(c)))
+				err = p.Interface.ExecuteHookCleanUp(ctx, execHook, api)
 				if err != nil {
 					cmd.Printf("app %q ExecuteHookCleanUp() error: %v", p.Path, err)
 				}
@@ -246,9 +245,8 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 
 	postCmd := cmd.PostRunE
 	cmd.PostRunE = func(cmd *cobra.Command, args []string) error {
-		// Get chain when the plugin runs inside an blockchain app
-		c, err := newChainWithHomeFlags(cmd)
-		if err != nil && !errors.Is(err, gomodule.ErrGoModNotFound) {
+		api, err := newAppClientAPI(cmd)
+		if err != nil {
 			return err
 		}
 
@@ -256,7 +254,7 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 		execHook := newExecutedHook(hook, cmd, args)
 
 		defer func() {
-			err := p.Interface.ExecuteHookCleanUp(ctx, execHook, plugin.NewClientAPI(plugin.WithChain(c)))
+			err := p.Interface.ExecuteHookCleanUp(ctx, execHook, api)
 			if err != nil {
 				cmd.Printf("app %q ExecuteHookCleanUp() error: %v", p.Path, err)
 			}
@@ -270,7 +268,7 @@ func linkPluginHook(rootCmd *cobra.Command, p *plugin.Plugin, hook *plugin.Hook)
 			}
 		}
 
-		err = p.Interface.ExecuteHookPost(ctx, execHook, plugin.NewClientAPI(plugin.WithChain(c)))
+		err = p.Interface.ExecuteHookPost(ctx, execHook, api)
 		if err != nil {
 			return errors.Errorf("app %q ExecuteHookPost() error : %w", p.Path, err)
 		}
@@ -333,9 +331,8 @@ func linkPluginCmd(rootCmd *cobra.Command, p *plugin.Plugin, pluginCmd *plugin.C
 		newCmd.RunE = func(cmd *cobra.Command, args []string) error {
 			ctx := cmd.Context()
 			return clictx.Do(ctx, func() error {
-				// Get chain when the plugin runs inside an blockchain app
-				c, err := newChainWithHomeFlags(cmd)
-				if err != nil && !errors.Is(err, gomodule.ErrGoModNotFound) {
+				api, err := newAppClientAPI(cmd)
+				if err != nil {
 					return err
 				}
 
@@ -348,7 +345,7 @@ func linkPluginCmd(rootCmd *cobra.Command, p *plugin.Plugin, pluginCmd *plugin.C
 					With:   p.With,
 				}
 				execCmd.ImportFlags(cmd)
-				err = p.Interface.Execute(ctx, execCmd, plugin.NewClientAPI(plugin.WithChain(c)))
+				err = p.Interface.Execute(ctx, execCmd, api)
 
 				// NOTE(tb): This pause gives enough time for go-plugin to sync the
 				// output from stdout/stderr of the plugin. Without that pause, this
@@ -424,18 +421,12 @@ If no path is specified all declared apps are updated.`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) == 0 {
 				// update all plugins
-				if err := plugin.Update(plugins...); err != nil {
-					return err
-				}
-				return nil
+				return plugin.Update(plugins...)
 			}
 			// find the plugin to update
 			for _, p := range plugins {
 				if p.HasPath(args[0]) {
-					if err := plugin.Update(p); err != nil {
-						return err
-					}
-					return nil
+					return plugin.Update(p)
 				}
 			}
 			return errors.Errorf("App %q not found", args[0])
@@ -700,6 +691,21 @@ func printPlugins(ctx context.Context, session *cliui.Session) error {
 		return errors.Errorf("error while printing apps: %w", err)
 	}
 	return nil
+}
+
+func newAppClientAPI(cmd *cobra.Command) (plugin.ClientAPI, error) {
+	// Get chain when the plugin runs inside an blockchain app
+	c, err := chain.NewWithHomeFlags(cmd)
+	if err != nil && !errors.Is(err, gomodule.ErrGoModNotFound) {
+		return nil, err
+	}
+
+	var options []plugin.APIOption
+	if c != nil {
+		options = append(options, plugin.WithChain(c))
+	}
+
+	return plugin.NewClientAPI(options...), nil
 }
 
 func flagSetPluginsGlobal() *flag.FlagSet {