+++ ignite/cmd/scaffold_module.go
@@ -12,6 +12,7 @@ import (
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
 	"github.com/ignite/cli/v29/ignite/pkg/placeholder"
 	"github.com/ignite/cli/v29/ignite/pkg/validation"
+	"github.com/ignite/cli/v29/ignite/pkg/xgenny"
 	"github.com/ignite/cli/v29/ignite/services/scaffolder"
 	modulecreate "github.com/ignite/cli/v29/ignite/templates/module/create"
 )
@@ -25,6 +26,7 @@ const (
 	flagDep                 = "dep"
 	flagIBC                 = "ibc"
 	flagParams              = "params"
+	flagModuleConfigs       = "module-configs"
 	flagIBCOrdering         = "ordering"
 	flagRequireRegistration = "require-registration"
 
@@ -115,6 +117,7 @@ params.
 	c.Flags().String(flagIBCOrdering, "none", "channel ordering of the IBC module [none|ordered|unordered]")
 	c.Flags().Bool(flagRequireRegistration, false, "fail if module can't be registered")
 	c.Flags().StringSlice(flagParams, []string{}, "add module parameters")
+	c.Flags().StringSlice(flagModuleConfigs, []string{}, "add module configs")
 
 	return c
 }
@@ -128,21 +131,12 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()
 
-	ibcModule, err := cmd.Flags().GetBool(flagIBC)
-	if err != nil {
-		return err
-	}
+	ibcModule, _ := cmd.Flags().GetBool(flagIBC)
+	ibcOrdering, _ := cmd.Flags().GetString(flagIBCOrdering)
+	requireRegistration, _ := cmd.Flags().GetBool(flagRequireRegistration)
+	params, _ := cmd.Flags().GetStringSlice(flagParams)
 
-	ibcOrdering, err := cmd.Flags().GetString(flagIBCOrdering)
-	if err != nil {
-		return err
-	}
-	requireRegistration, err := cmd.Flags().GetBool(flagRequireRegistration)
-	if err != nil {
-		return err
-	}
-
-	params, err := cmd.Flags().GetStringSlice(flagParams)
+	moduleConfigs, err := cmd.Flags().GetStringSlice(flagModuleConfigs)
 	if err != nil {
 		return err
 	}
@@ -154,6 +148,7 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 
 	options := []scaffolder.ModuleCreationOption{
 		scaffolder.WithParams(params),
+		scaffolder.WithModuleConfigs(moduleConfigs),
 	}
 
 	// Check if the module must be an IBC module
@@ -162,10 +157,7 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 	}
 
 	// Get module dependencies
-	dependencies, err := cmd.Flags().GetStringSlice(flagDep)
-	if err != nil {
-		return err
-	}
+	dependencies, _ := cmd.Flags().GetStringSlice(flagDep)
 	if len(dependencies) > 0 {
 		var deps []modulecreate.Dependency
 
@@ -204,7 +196,7 @@ func scaffoldModuleHandler(cmd *cobra.Command, args []string) error {
 			return err
 		}
 	} else {
-		modificationsStr, err := sourceModificationToString(sm)
+		modificationsStr, err := xgenny.SourceModificationToString(sm)
 		if err != nil {
 			return err
 		}