+++ ignite/services/scaffolder/module.go
@@ -105,16 +105,19 @@ var (
 
 // moduleCreationOptions holds options for creating a new module.
 type moduleCreationOptions struct {
-	// ibc true if the module is an ibc module
+	// ibc true if the module is an ibc module.
 	ibc bool
 
-	// params list of parameters
+	// params list of parameters.
 	params []string
 
-	// ibcChannelOrdering ibc channel ordering
+	// moduleConfigs list of module configs.
+	moduleConfigs []string
+
+	// ibcChannelOrdering ibc channel ordering.
 	ibcChannelOrdering string
 
-	// dependencies list of module dependencies
+	// dependencies list of module dependencies.
 	dependencies []modulecreate.Dependency
 }
 
@@ -132,6 +135,13 @@ func WithIBC() ModuleCreationOption {
 func WithParams(params []string) ModuleCreationOption {
 	return func(m *moduleCreationOptions) {
 		m.params = params
+	}
+}
+
+// WithModuleConfigs scaffolds a module with module configs.
+func WithModuleConfigs(moduleConfigs []string) ModuleCreationOption {
+	return func(m *moduleCreationOptions) {
+		m.moduleConfigs = moduleConfigs
 	}
 }
 
@@ -196,6 +206,12 @@ func (s Scaffolder) CreateModule(
 		return sm, err
 	}
 
+	// Parse configs with the associated type
+	configs, err := field.ParseFields(creationOpts.moduleConfigs, checkForbiddenTypeIndex)
+	if err != nil {
+		return sm, err
+	}
+
 	// Check dependencies
 	if err := checkDependencies(creationOpts.dependencies, s.path); err != nil {
 		return sm, err
@@ -205,6 +221,7 @@ func (s Scaffolder) CreateModule(
 		ModuleName:   moduleName,
 		ModulePath:   s.modpath.RawPath,
 		Params:       params,
+		Configs:      configs,
 		AppName:      s.modpath.Package,
 		AppPath:      s.path,
 		IsIBC:        creationOpts.ibc,